requirements.txt
python-telegram-bot==20.7
pyrogram==2.0.106
sqlalchemy==2.0.23
pyrogram-types
TgCrypto